Human avatar generator/requirements.txt
streamlit 
torch 
pillow 
fooocus